Fixed DAOs return types

+++ RevatureTrainingRoomPlanner/app/src/main/java/com/revature/revaturetrainingroomplanner/data/persistence/dao/BaseDAO.java
@@ -22,11 +22,11 @@ public interface BaseDAO<T> {
     LiveData<T> getByID(int id);
 
     @Update
-    long update(T... objects);
+    int update(T... objects);
 
     @Delete
-    long delete(T... objects);
+    int delete(T... objects);
 
-    long deleteAll();
+    int deleteAll();
 
 }
